requirements.txt
Flask==3.1.0
Flask-SocketIO==5.4.1
redis==5.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
eventlet==0.35.2
gevent==24.2.1
gevent-websocket==0.10.1